Adjustments to allow build of WWAudio on Linux

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWSaveLoad/persistfactory.h
@@ -50,6 +50,7 @@
 #include "chunkio.h"
 #include "wwdebug.h"
 #include "saveload.h"
+#include "persist.h"
 
 class PersistClass;
 
@@ -130,7 +131,8 @@ SimplePersistFactoryClass<T,CHUNKID>::Load(ChunkLoadClass & cload) const
 template<class T, int CHUNKID> void
 SimplePersistFactoryClass<T,CHUNKID>::Save(ChunkSaveClass & csave,PersistClass * obj) const 
 {
-	uint32 objptr = (uint32)obj;
+	// TODO: Merge upper and lower 32 bits of pointer on 64-bit systems
+	uint32 objptr = (uint32)(uintptr_t)obj;
 	csave.Begin_Chunk(SIMPLEFACTORY_CHUNKID_OBJPOINTER);
 	csave.Write(&objptr,sizeof(uint32));
 	csave.End_Chunk();
